Use deadlines for every read/write, enable TCP keepalive.
This is to reliably detect unresponsive servers.
See https://github.com/robustirc/robustirc/issues/115 for context.

This change also prevents a lockup in the bridge where the bridge
would hang in the sync.WaitGroup.Wait call because
(encoding/json).Decoder.Decode() would block on the Read until the
underlying connection times out on the TCP level (can take O(hours)).
This lockup is now impossible because the Read will return with an
error after 70 seconds.