Add -bridge_auth= flag to authenticate to RobustIRC networks
If specified, the file will be searched for a <network>:<secret-key>
line matching the network the bridge is about to connect to. If such an
entry is found, the X-Bridge-Auth header will contain the secret key.

You can create such a file using:

    touch /var/lib/bridge.auth
    chmod o-rwx /var/lib/bridge.auth
    echo "robustirc.net:$(openssl rand -base64 32)" >> /var/lib/bridge.auth

For more details, see https://github.com/robustirc/robustirc/issues/122