rework session deletion: make getMessages() close the channels
Before this change, Delete() closed the channels, but there was no protection in
place to ensure that getMessages() won’t write to the channel.

After this change, the only goroutine that ever writes to these channels is also
responsible for closing them.

Also remove s.deleted, which was not protected for access/modification from
concurrent goroutines (but happened to work in practice).