robustsession: handle non-HTTP transports containing slashes
Depending on the environment, RobustIRC can be deployed with additional layers
of transports, e.g. an HTTP-over-RPC mechanism.

If that transport’s target address contains slashes, it cannot be represented as
a HTTP URL hostname (http://<hostname>/<path>), as slash is the URL delimiter.

Instead, when a slash is detected as part of the target, we put the target into
the Host field/header (where a transport can extract it from) and create the
HTTP request only with a path instead of a full URL.

This should be a no-op for regular RobustIRC.